configuração padrão
