+++ lp-server/spec/lp-server.spec.ts
@@ -10,6 +10,14 @@ describe("O servidor", () => {
 
   afterAll(() => {server.closeServer()});
 
-  // add server tests here
+  it("inicialmente retorna um ranking vazio", () => {
+    return request.get(base_url + "estudoscomparativos")
+            .then(body => 
+               expect(body).toBe("[]")
+             )
+            .catch(e => 
+               expect(e).toEqual(null)
+             );
+  })
 
-})
+})
